// data/github/current/ui/packages/issue-create/__tests__/__generated__/CreateIssueDialogEntryTestQuery.graphql.ts
export type CreateIssueDialogEntryTestQuery$variables = {
  includeTemplates?: boolean | null | undefined;
  name: string;
  owner: string;
};
export type CreateIssueDialogEntryTestQuery$data = {
  readonly repository: {
    readonly " $fragmentSpreads": FragmentRefs<"CreateIssueDialog" | "RepositoryPickerRepository">;
  } | null | undefined;
};
export type CreateIssueDialogEntryTestQuery = {
  response: CreateIssueDialogEntryTestQuery$data;
  variables: CreateIssueDialogEntryTestQuery$variables;
};